% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{biocVersions}
\alias{biocVersions}
\title{Bioconductor version numbers}
\usage{
biocVersions()
}
\value{
A named \code{character} of length two with the
    \code{release} and \code{devel} version numbers (in that
    order).
}
\description{
This function returns the current release and devel Bioconductor
version numbers by parsing the
\code{"http://www.bioconductor.org/js/versions.js"} page.
}
\examples{
biocVersions()
}
\author{
Laurent Gatto
}
