% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribs.R
\name{contribs}
\alias{contribs}
\alias{ncontribs}
\alias{pkgrole}
\alias{pkgaut}
\alias{pkgcre}
\title{Getting package contributors}
\usage{
contribs(pkg)

ncontribs(pkg)

pkgrole(pkg, role.)

pkgaut(pkg)

pkgcre(pkg)
}
\arguments{
\item{pkg}{The name of the package.}

\item{role.}{The role to be searched for. See `?person` for
details.}
}
\value{
An instance of class `person` or the `character` with the
    author information and a warning (`contribs`) or an error
    (other functions) when `Authors@R` is not present in the
    package's `DESCRIPTION`. For `ncontribs` an `integer`.
}
\description{
These functions extract information about package contributors.
}
\examples{
contribs("MSnbase")
contribs("Biobase")
ncontribs("MSnbase")
pkgcre("MSnbase")
## same as
pkgrole("MSnbase", "cre")
pkgaut("rols")
## same as
pkgrole("rols", "aut")
}
\author{
Laurent Gatto
}
