% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pkg_dep_graph}
\alias{pkg_dep_graph}
\alias{plot_pkg_dep_graph}
\title{Build the package dependency graph}
\usage{
pkg_dep_graph(pkgs, which = c("BioCsoft", "BioCann", "BioCexp", "CRAN"))

plot_pkg_dep_graph(gr, sz = 20, fs = 50, pkgs, colour = "steelblue",
  edge_colour = "#00000040")
}
\arguments{
\item{pkgs}{The name of the package(s) to generate and plot the
dependency graph of. For plotting, it can be a list (each
element of the list being highlighted using different colours)
or a vector (all highlighted with the same colour) of package
names to be highlighted.}

\item{which}{What package repository to use to calculate
dependency graphs. One of `"BioCsoft"`(default, Bioconductor
software packages), `"BioCann"` (Bioconductor annotation
packages), `"BioCexp"` (Bioconductor experiment packages) or
`"CRAN"` (CRAN packages).}

\item{gr}{A `graphNEL` package dependency graph, as generated by
`pkg_dep_graph`.}

\item{sz}{The size of the nodes. Default is 20.}

\item{fs}{The font size. Default is 50.}

\item{colour}{A vector of colours to be used to highlight to
packages. The length of this argument must be of same length
as the `pkgs`.}

\item{edge_colour}{The colour of the edges. Default in
`"#00000040"`.}
}
\value{
An object of class `graphNEL`
}
\description{
This function takes a list of packages and build the their
dependency graph using all packages of a repository (for example
all Bioconductor software or CRAN packages). The packages of
interest and all their dependencies are used to create the graph.
}
\examples{
p <- c("MSnbase", "mzR")
g <- pkg_dep_graph(pkgs = p)
g
plot_pkg_dep_graph(g, fs = 25)
plot_pkg_dep_graph(g, pkgs = p, fs = 25)
plot_pkg_dep_graph(g, pkgs = list("MSnbase", "mzR"),
                   colour = c("steelblue", "yellow"),
                   fs = 25)
}
\author{
Laurent Gatto
}

