% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pkg_dep_graph}
\alias{pkg_dep_graph}
\title{Build the package dependency graph}
\usage{
pkg_dep_graph(pkgs, which = c("BioCsoft", "BioCann", "BioCexp", "CRAN"))
}
\arguments{
\item{pkgs}{The name of the package(s) to plot the dependency
graph of.}

\item{which}{What package repository to use to calculate
dependency graphs. One of `"BioCsoft"`(default, Bioconductor
software packages), `"BioCann"` (Bioconductor annotation
packages), `"BioCexp"` (Bioconductor experiment packages) or
`"CRAN"` (CRAN packages).}
}
\value{
An object of class `graphNEL`
}
\description{
This function takes a list of packages and build the their
dependency graph using all packages of a repository (for example
all Bioconductor software or CRAN packages). The packages of
interest and all their dependencies are used to create the graph.
}
\examples{
pp <- RforProteomics:::proteomicsPackages()
ppn <- rownames(pp)
ppn <- ppn[1:20]
g <- pkg_dep_graph(pkgs = ppn)
plot_dep_graph(g, pkgs = ppn, fs = "75")
}
\author{
Laurent Gatto
}
