% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocdownloads.R
\name{pkg_download_data}
\alias{pkg_download_data}
\title{Retrieve download data}
\usage{
pkg_download_data(pkg, by = c("month", "year"), ...)
}
\arguments{
\item{pkg}{A `character` with Bioconductor package names.}

\item{by}{Should the monthly downloads (`by = "month"`, default)
or the total per years (`by = "year"`) be downloaded.}

\item{...}{Additional arguments passed to
[utils::download.file()].}
}
\value{
A `tibble` with the download data.
}
\description{
Given one or more Bioconductor package names,
}
\examples{
pkg_download_data("Biobase")
pkg_download_data("Biobase", by = "year")
pkg_download_data("Biobase", by = "year", quiet = TRUE)
}
\author{
Laurent Gatto
}
