% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{biocpkgversion}
\alias{biocpkgversion}
\alias{githubpkgversion}
\title{Gets the package version numbers}
\usage{
biocpkgversion(pkg, which = c("release", "devel"), ...)

githubpkgversion(pkg, ...)
}
\arguments{
\item{pkg}{The name of the package(s) or, for \code{githubpkgversion},
the username and pkg formatted as \code{"user/package"}.}

\item{which}{Either of \code{"release"} (default) or \code{"devel"}.}

\item{...}{additional parameters passed to \code{sapply}, in particular
\code{simplify} (default is \code{TRUE}) and \code{USE.NAMES} (default is
\code{TRUE}).}
}
\value{
A \code{vector} of length equal to the length of \code{pkg}
containing the respective version numbers. By default, the
return value is named (\code{USE.NAMES = TRUE}) and simplified to a
\code{character} vector (\code{simplify = TRUE}).
}
\description{
A set of functions to retrieve the version number from packages
available in Bioconductor or on Github. The version numbers are
extracted from the Bioconductor package landing page and the
Github master branch, irrespective or the package's local
version of availability.
}
\examples{
biocpkgversion("MSnbase", "release")
biocpkgversion("MSnbase", USE.NAMES = FALSE)
biocpkgversion("MSnbase", "devel")
biocpkgversion(c("MSnbase", "pRoloc"))
biocpkgversion(c("MSnbase", "pRoloc"), USE.NAMES = FALSE)
githubpkgversion("lgatto/MSnbase")
githubpkgversion("lgatto/MSnbase", USE.NAMES = FALSE)
githubpkgversion(c("lgatto/MSnbase", "lgatto/pRoloc"))
}
\author{
Laurent Gatto
}
