% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{biocpkgversion}
\alias{biocpkgversion}
\alias{githubpkgversion}
\alias{cranpkgversion}
\title{Gets the package version numbers}
\usage{
biocpkgversion(pkg, which = c("release", "devel"), type = c("software",
  "experiment", "annotation"), ...)

githubpkgversion(pkg, user = NULL, ...)

cranpkgversion(pkg, ...)
}
\arguments{
\item{pkg}{The name of the package(s) or, for \code{githubpkgversion},
if \code{user} is \code{NULL} (default) the username and pkg formatted
as \code{"user/package"}.}

\item{which}{Either of \code{"release"} (default) or \code{"devel"}.}

\item{type}{One of \code{"software"}, \code{"experiment"} or \code{"annotation"},
defining the type of Bioconductor package is to be queried.}

\item{...}{additional parameters passed to \code{sapply}, in particular
\code{simplify} (default is \code{TRUE}) and \code{USE.NAMES} (default is
\code{TRUE}).}

\item{user}{the Github user or organisation where to find the
package repository. Default is `NULL` and ignored. Otherwise,
the packge name is constructed by concatenating to
`"user/pkg"`.}
}
\value{
A \code{vector} of length equal to the length of \code{pkg}
containing the respective version numbers. By default, the
return value is named (\code{USE.NAMES = TRUE}) and simplified to a
\code{character} vector (\code{simplify = TRUE}).
}
\description{
A set of functions to retrieve the version number from packages
available in Bioconductor or on Github. The version numbers are
extracted from the Bioconductor package landing page and the
Github master branch, irrespective or the package's local
version or availability.
}
\examples{
biocpkgversion("MSnbase", "release")
biocpkgversion("MSnbase", USE.NAMES = FALSE)
biocpkgversion("MSnbase", "devel")
biocpkgversion(c("MSnbase", "pRoloc"))
biocpkgversion(c("MSnbase", "pRoloc"), USE.NAMES = FALSE)
githubpkgversion("lgatto/MSnbase")
githubpkgversion("MSnbase", "lgatto")
githubpkgversion("lgatto/MSnbase", USE.NAMES = FALSE)
githubpkgversion(c("lgatto/MSnbase", "lgatto/pRoloc"))
cranpkgversion(c("sequences", "MALDIquant"))
}
\author{
Laurent Gatto
}
